requirements.txt
torch
torchvision
transformers
datasets
pillow
pandas
scikit-learn
tqdm